remove lastID hack for incomplete datablocks
The Fortran version processes input files with EDBL (end of block statement) immediately followed by a dataset (initiated by DATA), which leads to double counting of data. This commit removes a hack to reproduce the Fortran behavior in the case of incomplete datablocks, hence the results of the Fortran and Python GMAP version will differ in this scenario.