refactor: use (de)sanitize_datablock funs as a check
The Python GMAP version will use a new format for the
storage of the datablocks. In order to ensure that the new
representation is equivalent, the datablocks given in
the old format are converted to the new one and back.
The final result is used in the fitting and it is
checked that the output is indeed equivalent.