refactor: split code into packages
Move all essential functions for GMA calculations
into the package gmapi (as a portmanteau of GMA and API).
This package contains two subpackages mappings and legacy.
The legacy package contains all functions to read and
write data in legacy format, and to convert the legacy
database to a JSON database.