replace legacy basic mapping in absolute_ratio_map
Replace the legacy basic mapping used in the class
CrossSectionAbsoluteRatioMap which required
the energies of the target mesh to be a subset of
the energies of the source mesh by a more versatile
basic mapping. Energies of source and target mesh
don't need to exactly match anymore, but linear
interpolation is employed.